Move executable to bin folder
