+++ eva_sub_cli/docker_validator.py
@@ -14,8 +14,8 @@ from ebi_eva_common_pyutils.logger import logging_config
 logger = logging_config.get_logger(__name__)
 
 docker_path = 'docker'
-container_image = 'ebivariation/eva-sub-eva_sub_cli'
-container_tag = 'v0.0.1.dev0'
+container_image = 'ebivariation/eva-sub-cli'
+container_tag = 'v0.0.1.dev1'
 container_validation_dir = '/opt/vcf_validation'
 container_validation_output_dir = '/opt/vcf_validation/vcf_validation_output'
 container_etc_dir = '/opt/eva_sub_cli/etc'
@@ -92,6 +92,7 @@ class DockerValidator(Reporter):
 
             docker_cmd = self.get_docker_validation_cmd()
             # start validation
+            print(docker_cmd)
             # FIXME: If nextflow fails in the docker exec still exit with error code 0
             run_command_with_output("Run Validation using Nextflow", docker_cmd)
             # copy validation result to user host
